Improve window detection, login timing, and input speed
- Add tastytrade login screen detection via title regex and is_login_screen()
- Fix xdotool focus_window to use click-to-focus instead of unreliable --sync
- Wire login_ready_delay from platform config (was hardcoded 1s sleep)
- Add timing instrumentation for screenshot, OCR, and form fill stages
- Prefer clipboard paste over typewrite for faster credential entry
- Crop screenshots to window geometry for more accurate OCR
- Add grayscale preprocessing before tesseract for speed
- Reduce polling interval and inter-keystroke delays
- Add secretstorage-to-keyring one-time migration helper
- Tighten TWS is_login_screen() to exact "Login" title match

+++ platforms/base.py
@@ -25,7 +25,7 @@ class PlatformConfig:
     wm_class: str = ""
     title_pattern: str = ""
     window_timeout: int = 120
-    login_ready_delay: int = 8
+    login_ready_delay: float = 8
     field_order: list[str] = field(default_factory=lambda: ["username", "password"])
     input_strategy: str = "auto"
     extra: dict[str, Any] = field(default_factory=dict)
